+++ TunnelKit/Sources/Protocols/OpenVPN/OpenVPNSession.swift
@@ -3,7 +3,7 @@
 //  TunnelKit
 //
 //  Created by Davide De Rosa on 2/3/17.
-//  Copyright (c) 2019 Davide De Rosa. All rights reserved.
+//  Copyright (c) 2020 Davide De Rosa. All rights reserved.
 //
 //  https://github.com/passepartoutvpn
 //
@@ -55,10 +55,11 @@ public protocol OpenVPNSessionDelegate: class {
     /**
      Called after stopping a session.
      
+     - Parameter error: An optional `Error` being the reason of the stop.
      - Parameter shouldReconnect: When `true`, the session can/should be restarted. Usually because the stop reason was recoverable.
      - Seealso: `OpenVPNSession.reconnect(...)`
      */
-    func sessionDidStop(_: OpenVPNSession, shouldReconnect: Bool)
+    func sessionDidStop(_: OpenVPNSession, withError error: Error?, shouldReconnect: Bool)
 }
 
 /// Provides methods to set up and maintain an OpenVPN session.
@@ -160,7 +161,7 @@ public class OpenVPNSession: Session {
 
     private var lastPing: BidirectionalState<Date>
     
-    private var isStopping: Bool
+    private(set) var isStopping: Bool
     
     /// The optional reason why the session stopped.
     public private(set) var stopError: Error?
@@ -407,6 +408,8 @@ public class OpenVPNSession: Session {
             }
             if let error = error {
                 log.error("Failed LINK read: \(error)")
+                
+                // XXX: why isn't the tunnel shutting down at this point?
                 return
             }
             
@@ -428,7 +431,7 @@ public class OpenVPNSession: Session {
             }
 
             if let packets = newPackets, !packets.isEmpty {
-//                log.verbose("Received \(packets.count) packets from \(self.tunnelName)")
+//                log.verbose("Received \(packets.count) packets from TUN")
                 self?.receiveTunnel(packets: packets)
             }
         }
@@ -437,6 +440,7 @@ public class OpenVPNSession: Session {
     // Ruby: recv_link
     private func receiveLink(packets: [Data]) {
         guard shouldHandlePackets() else {
+            log.warning("Discarding \(packets.count) LINK packets (should not handle)")
             return
         }
         
@@ -472,7 +476,7 @@ public class OpenVPNSession: Session {
                 guard let _ = keys[key] else {
                     log.warning("Key with id \(key) not found")
 //                    deferStop(.shutdown, OpenVPNError.badKey)
-                    return
+                    continue // JK: This used to be return, but we'd see connections that would stay in Connecting… state forever
                 }
 
                 // XXX: improve with array reference
@@ -536,46 +540,46 @@ public class OpenVPNSession: Session {
     // Ruby: recv_tun
     private func receiveTunnel(packets: [Data]) {
         guard shouldHandlePackets() else {
+            log.warning("Discarding \(packets.count) TUN packets (should not handle)")
             return
         }
         sendDataPackets(packets)
-        lastPing.outbound = Date()
     }
     
     // Ruby: ping
     private func ping() {
-        guard (currentKey?.controlState == .connected) else {
+        guard currentKey?.controlState == .connected else {
             return
         }
         
         let now = Date()
-        guard (now.timeIntervalSince(lastPing.inbound) <= keepAliveTimeout) else {
+        guard now.timeIntervalSince(lastPing.inbound) <= keepAliveTimeout else {
             deferStop(.shutdown, OpenVPNError.pingTimeout)
             return
         }
 
-        // postpone ping if elapsed less than keep-alive
-        if let interval = keepAliveInterval {
-            let elapsed = now.timeIntervalSince(lastPing.outbound)
-            guard (elapsed >= interval) else {
-                scheduleNextPing(elapsed: elapsed)
-                return
-            }
+        // is keep-alive enabled?
+        if let _ = keepAliveInterval {
+            log.debug("Send ping")
+            sendDataPackets([OpenVPN.DataPacket.pingString])
+            lastPing.outbound = Date()
         }
 
-        log.debug("Send ping")
-        sendDataPackets([OpenVPN.DataPacket.pingString])
-        lastPing.outbound = Date()
-
+        // schedule even just to check for ping timeout
         scheduleNextPing()
     }
     
-    private func scheduleNextPing(elapsed: TimeInterval = 0.0) {
-        guard let interval = keepAliveInterval else {
-            return
+    private func scheduleNextPing() {
+        let interval: TimeInterval
+        if let keepAliveInterval = keepAliveInterval {
+            interval = keepAliveInterval
+            log.verbose("Schedule ping after \(interval) seconds")
+        } else {
+            interval = CoreConfiguration.OpenVPN.pingTimeoutCheckInterval
+            log.verbose("Schedule ping timeout check after \(interval) seconds")
         }
-        let remaining = min(interval, interval - elapsed)
-        queue.asyncAfter(deadline: .now() + remaining) { [weak self] in
+        queue.asyncAfter(deadline: .now() + interval) { [weak self] in
+            log.verbose("Running ping block")
             self?.ping()
         }
     }
@@ -742,7 +746,7 @@ public class OpenVPNSession: Session {
     
     private func completeConnection() {
         setupEncryption()
-        authenticator = nil
+        authenticator?.reset()
         negotiationKey.controlState = .connected
         connectedDate = Date()
         transitionKeys()
@@ -784,7 +788,9 @@ public class OpenVPNSession: Session {
                 caPath: caURL.path,
                 clientCertificatePath: (configuration.clientCertificate != nil) ? clientCertificateURL.path : nil,
                 clientKeyPath: (configuration.clientKey != nil) ? clientKeyURL.path : nil,
-                checksEKU: configuration.checksEKU ?? false
+                checksEKU: configuration.checksEKU ?? false,
+                checksSANHost: configuration.checksSANHost ?? false,
+                hostname: configuration.sanHost
             )
             if let tlsSecurityLevel = configuration.tlsSecurityLevel {
                 tls.securityLevel = tlsSecurityLevel
@@ -904,6 +910,11 @@ public class OpenVPNSession: Session {
 
     // Ruby: handle_ctrl_msg
     private func handleControlMessage(_ message: String) {
+        if CoreConfiguration.logsSensitiveData {
+            log.debug("Received control message: \"\(message)\"")
+        }
+
+        // disconnect on authentication failure
         guard !message.hasPrefix("AUTH_FAILED") else {
 
             // XXX: retry without client options
@@ -918,14 +929,18 @@ public class OpenVPNSession: Session {
             return
         }
         
-        guard (negotiationKey.controlState == .preIfConfig) else {
+        // disconnect on remote server restart (--explicit-exit-notify)
+        guard !message.hasPrefix("RESTART") else {
+            log.debug("Disconnecting due to server shutdown")
+            deferStop(.shutdown, OpenVPNError.serverShutdown)
+            return
+        }
+        
+        // handle authentication from now on
+        guard negotiationKey.controlState == .preIfConfig else {
             return
         }
 
-        if CoreConfiguration.logsSensitiveData {
-            log.debug("Received control message: \"\(message)\"")
-        }
-        
         let completeMessage: String
         if let continuated = continuatedPushReplyMessage {
             completeMessage = "\(continuated),\(message)"
@@ -1221,7 +1236,7 @@ public class OpenVPNSession: Session {
     // MARK: Stop
     
     private func shouldHandlePackets() -> Bool {
-        return (!isStopping && !keys.isEmpty)
+        return !isStopping && !keys.isEmpty
     }
     
     private func deferStop(_ method: StopMethod, _ error: Error?) {
@@ -1267,7 +1282,7 @@ public class OpenVPNSession: Session {
             log.info("Trigger shutdown on request")
         }
         stopError = error
-        delegate?.sessionDidStop(self, shouldReconnect: false)
+        delegate?.sessionDidStop(self, withError: error, shouldReconnect: false)
     }
     
     private func doReconnect(error: Error?) {
@@ -1277,6 +1292,6 @@ public class OpenVPNSession: Session {
             log.info("Trigger reconnection on request")
         }
         stopError = error
-        delegate?.sessionDidStop(self, shouldReconnect: true)
+        delegate?.sessionDidStop(self, withError: error, shouldReconnect: true)
     }
 }